chore: move provider and providers files to new directory structure
