--- client/src/app/App.css
@@ -1,1 +1,0 @@
-
