+++ client/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "@/index.css";
 import App from "@/app/App";
-import { Providers } from "@/lib/provider/providers";
+import { Providers } from "@/app/providers";
 
 const container = document.getElementById("root");
